+++ src/measurement_kit/cmdline.hpp
@@ -16,6 +16,7 @@ using namespace mk::nettests;
 #define MK_CMDLINE_SUBCOMMANDS                                                 \
     XX(dns_injection)                                                          \
     XX(http_invalid_request_line)                                              \
+    XX(meek_fronted_requests)                                                  \
     XX(multi_ndt)                                                              \
     XX(ndt)                                                                    \
     XX(tcp_connect)                                                            \
